updated to work on CentOS 5
git-svn-id: https://crosswire.org/svn/sword/trunk@2655 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ src/mgr/curlhttpt.cpp
@@ -231,7 +231,7 @@ vector<struct DirEntry> CURLHTTPTransport::getDirList(const char *dirURL) {
 				}
 				struct DirEntry i;
 				i.name = possibleName;
-				i.size = fSize;
+				i.size = (long unsigned int)fSize;
 				i.isDirectory = (possibleName[possibleNameLength-1] == '/');
 				dirList.push_back(i);
 				pBuf = pBufRes;
